+++ src/app/epics/skrivemodusEpics.tsx
@@ -2,7 +2,7 @@ import { AppEpic } from '../store/configureStore';
 import { catchError, filter, map, switchMap } from 'rxjs/operators';
 import { isActionOf } from 'typesafe-actions';
 import { from, of } from 'rxjs';
-import { WeblogicActions } from '../actions/weblogic';
+import { WeblogicActions } from '../actions/skrivemodus';
 import { pingWeblogic } from '../api/api';
 import { MeldekortActions } from '../actions/meldekort';
 import { combineEpics } from 'redux-observable';
