+++ src/app/mock/responses/skrivemodus.json
