+++ src/app/sider/tidligereMeldekort/detaljer/detaljer.tsx
@@ -31,7 +31,7 @@ import classNames from 'classnames';
 import { AktivtMeldekortActions } from '../../../actions/aktivtMeldekort';
 import { HistoriskeMeldekortState } from '../../../reducers/historiskeMeldekortReducer';
 import { WeblogicPing } from '../../../types/weblogic';
-import { WeblogicActions } from '../../../actions/weblogic';
+import { WeblogicActions } from '../../../actions/skrivemodus';
 import { finnTypeYtelsePostfix } from '../../../utils/teksterUtil';
 import { downloadMessagesAndDispatch } from '../../../utils/intlUtil';
